reorganise resource configs
