+++ dag/src/dag/defs/configs.py
@@ -1,5 +1,14 @@
 import dagster as dg
+from typing import Annotated, Literal
 
+import dagster as dg
+from neural_optimiser.calculators import FAIRChemCalculator, MACECalculator, MMFF94Calculator
+from neural_optimiser.calculators.base import Calculator
+from neural_optimiser.optimisers import BFGS
+from neural_optimiser.optimisers.base import Optimiser
+from pydantic import Field
+
+# ---------- Assets Configs ----------
 
 class InputConfig(dg.Config):
     """Config for preprocess_data -> to_mols_dict()."""
@@ -46,3 +55,72 @@ class OutputConfig(dg.Config):
     molecule_attr: str | None = None
     id_col_name: str | None = None
     mol_col_name: str | None = None
+
+# ---------- Resources Configs ----------
+
+class MACEConfig(dg.Config):
+    kind: Literal["mace"] = "mace"
+    model_paths: str
+    device: str = "cpu"
+    default_dtype: Literal["float32", "float64"] = "float32"
+
+    def build(self) -> Calculator:
+        return MACECalculator(
+            model_paths=self.model_paths, device=self.device, default_dtype=self.default_dtype
+        )
+
+
+class MMFF94Config(dg.Config):
+    kind: Literal["mmff94"] = "mmff94"
+
+    def build(self) -> Calculator:
+        return MMFF94Calculator()
+
+
+class FAIRChemConfig(dg.Config):
+    kind: Literal["fairchem"] = "fairchem"
+    model_paths: str
+    device: str = "cpu"
+    task_name: str = "omol"
+    default_dtype: Literal["float32", "float64"] = "float32"
+
+    def build(self) -> Calculator:
+        return FAIRChemCalculator(
+            model_paths=self.model_paths,
+            device=self.device,
+            task_name=self.task_name,
+            default_dtype=self.default_dtype,
+        )
+
+
+CalculatorConfig: dg.Config = Annotated[
+    MACEConfig | MMFF94Config | FAIRChemConfig, Field(discriminator="kind")
+]
+
+
+class BFGSConfig(dg.Config):
+    kind: Literal["bfgs"] = "bfgs"
+    max_step: float = 0.04
+    steps: int = 250
+    fmax: float | None = None
+    fexit: float | None = None
+
+    def build(self) -> Optimiser:
+        return BFGS(max_step=self.max_step, steps=self.steps, fmax=self.fmax, fexit=self.fexit)
+
+
+class _BaseOptimiserConfig(dg.Config):
+    """Placeholder so OptimiserConfig is a discriminated union (i.e. renders as a selector,
+    uniform with CalculatorConfig). The base Optimiser is abstract, so this is never selected;
+    it exists only until a second concrete optimiser (e.g. LBFGS) is added."""
+
+    kind: Literal["base"] = "base"
+
+    def build(self) -> Optimiser:
+        raise NotImplementedError(
+            "Select a concrete optimiser (e.g. bfgs); 'base' is a placeholder."
+        )
+
+
+OptimiserConfig: dg.Config = Annotated[BFGSConfig | _BaseOptimiserConfig, Field(discriminator="kind")]
+
